server/requirements.txt
fastapi==0.70.0
cmyui==1.9.3
uvicorn[standard]
databases[mysql]
starlette
python-jose[cryptography]
bcrypt
hashlib
cmyui
python-multipart